Add a faster way to pick ngrams.
Instead of looping through all the possible ngrams in a list function
and picking one randomly, issue two requests, one to get the number of
rows, and a second to pick one by skipping a random number of rows.
Also choose whether to sort descending or not to skip less rows if one
in the second half of the list is picked.

With this method, on average, 1/4 as many rows will have to be iterated
through. However, it requires an additional request, and timing will not
be consistent.